NEUX: Fix bug that caused drawing entries in NListView without border.
The problem is that parent()->height() returns height of the
whole widget (so, plenty of space) and algorith computed the
height of the entry based on that: height / (height / fontHeight * (5/3)) =>
height * fontHeight * (5/3) / height => fontHeight * (5/3).
Pretty dumb, right?
But worse, sometimes parent()->height() returns small number (it is called
earlier in these cases (when setCurrentRow() called), so looks like something
is not updated yet) and we compute in integer numbers! So we get with
default size of font something like: height / 1, it is only lucky you
don't use bigger fonts to get something like: height / 0 => bang!